% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{as_jamstack}
\alias{as_jamstack}
\title{Generate jamstack data structure}
\usage{
as_jamstack(pkg = ".", override = list())
}
\arguments{
\item{pkg}{Path to package.}

\item{override}{An optional named list used to temporarily override
values in \verb{_jamstack.yml}}
}
\description{
You will generally not need to use this unless you need a custom site
design and you're writing your own equivalent of \code{\link[=build_site]{build_site()}}.
}
