% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamstackingup-package.R
\docType{package}
\name{jamstackingup-package}
\alias{jamstackingup-package}
\title{Creating Jamstack Markdown Package Documentation Websites}
\description{
By converting the package documentation files into markdown files and generating pages using page templates along the respective routes.
}
\seealso{
\code{knitr}
}
\author{
Saiem Gilani \href{mailto:saiem.gilani@gmail.com}{saiem.gilani@gmail.com}
}
\keyword{Rd}
\keyword{help}
\keyword{markdown}
