% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_jamstackup_site.R
\name{use_jamstackup_site}
\alias{use_jamstackup_site}
\title{Create jamstack site folder}
\usage{
use_jamstackup_site(
  pkg_dir = ".",
  site_dir = "website",
  template = "nextjs-ts"
)
}
\arguments{
\item{pkg_dir}{Directory of package}

\item{site_dir}{Directory to clone the website template in}

\item{template}{site template (default: "nextjs-ts")}
}
\description{
Shell script to build the website directory
}
