% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen2markdown.R
\name{roxygen2markdown}
\alias{roxygen2markdown}
\title{Rd file to markdown}
\usage{
roxygen2markdown(
  rdfile = "../wehoop-dev/wehoop/man/espn_wbb_player_box.Rd",
  outfile = "md/espn_wbb_player_box.md",
  append = FALSE
)
}
\arguments{
\item{rdfile}{Filepath to an .Rd file or an \code{Rd} object to parse.}

\item{outfile}{Filepath to output file (markdown file).}

\item{append}{If outfile exists, append to existing content.}
}
\value{
Parsed Rd as named list
}
\description{
This function converts an Rd file into markdown format.
}
\examples{
## give a markdown source file
rdfile = "~/git/MyPackage/man/myfun.Rd"
## specify, where markdown shall be stored
outfile = "/var/www/html/R_Web_app/md/myfun.md"
## create markdown
## roxygen2markdown(rdfile = rdfile, outfile = outfile)
}
